Config Blocks

Each block in the config file is structured as:

begin BLOCK_LABEL
    ...
end

Tags placed outside their designated blocks are ignored. The following blocks are available:

Atom tags

Hamster.get_alpha — Method

alpha=0.7*Z

The alpha::Float value determines how rapid the orbital overlap for a specific ion type falls off with distance. Defaults to 70% of the core charge.

Hamster.get_n — Method

n=n_period

The n::Int value determines the order of the polynomial that is used to model the distance dependence the orbital overlap for a specific ion type. Defaults to the period the atom species belongs to.

Hamster.get_nnaxes — Method

nnaxes=false

If nnaxes::Bool=true, the orbital axes are rotated along the connecting vectors with the nearest neighbors of the respective orbitals. The number of nearest neighbors depends on the number of orbitals.

Supercell

Hamster.get_Nconf — Method

Nconf=10

The Nconf tag sets the number of samples that are sampled from the total number of configuration.

Hamster.get_Nconf_max — Method

Nconf_max=1

The Nconf_max tag sets the maximum index that can be sampled from the total number of configurations.

Hamster.get_Nconf_min — Method

Nconf_min=10

The Nconf_min tag sets the minimum index that can be sampled from the total number of configurations.

Hamster.get_config_inds — Method

config_inds=none

The config_inds tag sets a file from which configuration indices are read. By default, indices are not read from file. It is also possible to provide a list of integers directly.

Hamster.get_nbatch — Method

nbatch=1

The nbatch tag detemines into how many batches the training structures are split for stochastic gradient optimization.

Hamster.get_sc_poscar — Method

sc_poscar=SC_POSCAR

The sc_poscar tag sets the path to the POSCAR file used for the supercell.

Hamster.get_val_config_inds — Method

config_inds=none

The config_inds tag sets a file from which configuration indices are read. By default, indices are not read from file. It is also possible to provide a list of integers directly.

Hamster.get_xdatcar — Method

xdatcar=XDATCAR

The xdatcar tag sets the path to the XDATCAR file.

ML

Hamster.get_apply_distance_distortion — Method

apply_distance_distortion=false

If true, distortions are applied to atomic positions in calculating atomic distance in ML descriptor.

Hamster.get_apply_distortion — Method

apply_distortion=false

If true, distortions are applied to atomic positions in calculating descriptor values.

Hamster.get_env_scale — Method

env_scale=1.

Sets the scaling factor that is multiplied with the environmental descriptor.

Hamster.get_ml_filename — Method

filename=ml_params

The filename tag sets the name for the parameter file of the ML model.

Hamster.get_ml_init_params — Method

init_params=zeros

The init_params tag determines how the parameters of the ML model are initialized.

Hamster.get_ml_model — Method

ml_model=false

The ml_model tag switches on the use of an ML model in the effective Hamiltonian.

Hamster.get_ml_ncluster — Method

ncluster=1

Sets the number of clusters to be used in the kmeans clustering.

Hamster.get_ml_npoints — Method

npoints=1

Sets the number of data points that are sampled in total.

Hamster.get_ml_rcut — Method

ml_rcut=rcut

Sets the cut-off radius for the ML model. Defaults to the same cut-off radius as the TB model.

Hamster.get_ml_sampling — Method

sampling=random

The sampling tag determines how points are selected from each cluster. Defaults to "random".

Hamster.get_ml_update — Method

update=true

The update tag switches on/off optimization of ML parameters.

Hamster.get_sim_params — Method

sim_params=0.1

Sets the parameter for the similarity function of the kernel model.

Hamster.get_strc_scale — Method

strc_scale=1.

Sets the scaling factor that is multiplied with the structural descriptor entries.

SOC

Hamster.get_soc — Method

soc=false

A model to account for spin-orbit coupling (SOC) is constructed if a block SOC is present in the config file.

Hamster.get_soc_init_params — Method

init_params=zeros

The init_params tag determines how SOC parameters are initialized.

Hamster.get_update_soc — Method

update_soc=true

The update_soc tag determines whether the SOC parameters are udpated.

Optim

Hamster.get_band_weights — Method

wE=ones

The wE tag sets the weight of each energy band for the calculation of the loss. Individual weights can also be set with, e.g., wE_3 = 2.

Hamster.get_bandmin — Method

bandmin=1

The bandmin tag sets the index of the lowest band that is included in the fitting.

Hamster.get_barrier — Method

barrier=0.

The barrier parameter determines at which magnitude the regularization kicks in.

Hamster.get_eig_fit — Method

eig_fit=true (false if hr_fit)

The eig_fit tag switches on fitting the effective Hamiltonian to eigenvalue data. If both hr_fit and eig_fit are true, the model is first fit to Hamiltonian data and then to eigenvalue data.

Hamster.get_eig_val — Method

eig_val=eig_fit

The eig_val tag switches on validating the effective Hamiltonian model with eigenvalue data.

Hamster.get_hr_fit — Method

hr_fit=false

The hr_fit tag switches on fitting the effective Hamiltonian model to Hamiltonian data.

Hamster.get_kpoint_weights — Method

wk=ones

The wk tag sets the weight of each kpoint for the calculation of the loss. Individual weights can also be set with, e.g., wk_3 = 2.

Hamster.get_lambda — Method

lambda=0.

The lambda parameter determines the regularization constant.

Hamster.get_loss — Method

loss=MAE

The loss tag sets the loss function to be used for the optimization.

Hamster.get_lr — Method

lr=0.1

The lr parameter defines the learning rate for the gradient descent parameter optimization.

Hamster.get_lr_min — Method

lr_min=0.1

The lr_min parameter defines the final learning rate when applying learning rate decay. Learning rate decay (using a cosine function) is automatically applied when lr_min is set to a value unequal lr.

Hamster.get_lreg — Method

lreg=2

The lreg tag determines which norm is used to calculate the regularization, e.g., L2 for lreg=2.

Hamster.get_niter — Method

niter=1

The niter parameter sets the maximum number of iterations (through the training data set).

Hamster.get_printeachbatch — Method

printeachbatch=false (true if verbosity > 1)

The printeachbatch tag switches on printing profiler information at every batch.

Hamster.get_printeachiter — Method

printeachiter=1

The printeachiter tag determines at what iteration interval profiler information is printed.

Hamster.get_train_data — Method

train_data=EIGENVAL

The train_data tag sets the path (filename) that contains the training data.

Hamster.get_train_mode — Method

train_mode=PC

The train_mode flag switches between different optimization modes (PC, MD, mixed, multi; not case sensitive).

Hamster.get_update_tb — Method

update_tb=haskey(conf, "Optimizer")

The update_tb tag determines whether/which parameters of the TB model are updated during the optimization. Defaults to true if an Optimizer block is provided.

Hamster.get_val_bandmin — Method

val_bandmin=1

The bandmin tag sets the index of the lowest band that is included in the validation.

Hamster.get_val_data — Method

val_data=EIGENVAL

The val_data tag sets the path (filename) that contains the validation data.

Hamster.get_val_mode — Method

val_mode=PC

The val_mode flag switches between different modes for model validation (PC, MD, mixed, multi; not case sensitive).

Hamster.get_val_ratio — Method

val_ratio=0.2 (if validate; 0 else)

The val_ratio tag sets the ratio between the training set size and the validation set size.

Hamster.get_val_weights — Method

val_weights=false

If val_weights is true, the same weights are used for the validation loss as for the training loss.

Hamster.get_valeachiter — Method

valeachiter=1

The valeachiter tag determines at what iteration interval a validation step is performed.

Hamster.get_validate — Method

validate=false (true if `val_mode` is set)

The validate tag switches on model validation.

HyperOpt

Hamster.get_hyperopt_lowerbounds — Method

lowerbounds=[0]

The lowerbounds::Vector{Float} define the lower bounds of the search space.

Hamster.get_hyperopt_mode — Method

mode = random

The mode::String tag specifies the strategy for generating new hyperparameter candidates. The default is random.

Possible options:

random: random search (default)

grid: grid search

tpe: Tree-structured Parzen Estimator

Hamster.get_hyperopt_niter — Method

niter=10

The niter::Int tags sets the maximum number of iterations in the hyperparameter optimization.

Hamster.get_hyperopt_params — Method

params=[""]

The params tag defines the hyperparameters to be optimized. For parameters belonging to a specific block, use the syntax block_param.

Hamster.get_hyperopt_stepsizes — Method

stepsizes=[1e-5]

The stepsizes::Vector{Float} defines the minimum difference between two points that are sampled as trial hyperparameters.

Hamster.get_hyperopt_upperbounds — Method

upperbounds=[0]

The upperbounds::Vector{Float64} define the upper bounds of the search space.