General Options

The general options are placed inside an Options block:

begin Options
    ...
end

Hamster.get_diag_method — Method

diag_method=shift-invert

The diag_method tag sets the method to be used for calculating eigenvalues when sp_diag=true, ignored otherwise.

Possible options:

shift-invert (default): call eigs function from Arpack.

krylov-schur: call eigsolve function from KrylovKit.

Hamster.get_eig_target — Method

eig_target=0.

The eig_target::Float tag sets the target energy when using a sparse eigensolver.

Hamster.get_init_params — Method

init_params = ones

The init_params tag determines how the tight-binding parameters are initialized.

Possible options:

filename.dat: initialize from a file named filename.dat.

ones: initialize all parameters to 1 (default).

zeros: initialize all parameters to 0.

rand: initialize parameters randomly between 0 and 1.

Hamster.get_kpoints_file — Method

kpoints=gamma

The kpoints::String tag specifies the file or method used to define the set of k-points.

Possible options:

gamma: use only the Gamma point (default).

EIGENVAL: read k-points from a VASP EIGENVAL file.

filename.h5: read k-points from the "k-points" field of an HDF5 file.

Hamster.get_neig — Method

neig=6 (only if sp_diag=true)

The neig::Int tag sets the number of eigenvalues that are calculated when using a sparse eigensolver.

Hamster.get_nhamster — Method

nhamster=1

The nhamster tag sets the number of Hamster processes to be spawned for parallel tasks.

Hamster.get_nthreads_bands — Method

nthreads_bands=JULIANUMTHREADS

The nthreads_bands::Int tag sets the number of tasks to work on energy bands simultaneously.

Hamster.get_nthreads_blas — Method

nthreads_blas=1

The nthreads_blas::Int tag sets the number of threads used by the BLAS library.

Hamster.get_nthreads_kpoints — Method

nthreads_kpoints=JULIANUMTHREADS

The nthreads_kpoints::Int tag sets the number of tasks to work on kpoints simultaneously.

Hamster.get_save_vecs — Method

save_vecs=false

The save_vecs::Bool tag determines whether the eigenvectors are written to a file.

Hamster.get_skip_diag — Method

skip_diag=false

If skip_diag is set to true, no eigenvalues are computed for the Hamiltonian.

Hamster.get_system — Method

system=unknown

The system tag gives a name to the system under study.

Hamster.get_verbosity — Method

verbosity=1

The verbosity parameter controls the amount of output printed to the console or written to output files. A value of 0 disables most print statements.

Possible options:

0: minimal output; most print statements are suppressed.

1: normal output volume (default).

2: increased output; additional information is printed.

3: maximum output; includes detailed debug information.

Hamster.get_write_hk — Method

write_hk=false

The write_hk::Bool tag determines whether the Hamiltonians in k-space are written to a file.

Hamster.set_seed! — Method

seed=none

The seed tag can be used to set a custom seed for RNG.

Hamster.get_Rmax — Method

Rmax=1

The parameter Rmax sets the maximum magnitude of a translation vector used for periodic boundaty conditions. Note that this is determined automatically if rcut is set.

Hamster.get_grid_size — Method

grid_size=rcut

The parameter grid_size determines the size of the cubes that are used to divide the simulation cell.

Hamster.get_poscar — Method

poscar=POSCAR

The parameter poscar sets the path to the POSCAR VASP file that defines the base system.

Hamster.get_rcut — Method

rcut=7.

The parameter rcut sets the cut-off radius beyond which all interactions are neglected. A cut-off radius of zero means all interactions are considered.

Hamster.get_rcut_tol — Method

rcut_tol=1.

The parameter rcut_tol sets a tolerance for the cutoff function.

Hamster.get_interpolate_rllm — Method

interpolate_rllm=true

The interpolate_rllm::Bool tag switches on interpolation of the distance dependence of overlaps.

Hamster.get_itp_Ninit — Method

itp_Ninit=20

The itp_Ninit::Int tag sets the number of initial points used for the adaptive interpolation.

Hamster.get_itp_Nmax — Method

itp_Nmax=1000

The itp_Nmax::Int tag sets the maximum number of points used for the adaptive interpolation.

Hamster.get_itp_tol — Method

itp_tol=1e-5

The itp_tol::Float tag sets the numerical tolerance for the adaptive interpolation.

Hamster.get_itp_xmax — Method

itp_xmax=rcut+rcut_tol+1

The itp_xmax::Float tag sets the maximal x value used for the adaptive interpolation.

Hamster.get_itp_xmin — Method

itp_xmin=0.

The itp_xmin::Float tag sets the minimal x value used for the adaptive interpolation.

Hamster.get_load_rllm — Method

load_rllm=false

The load_rllm::Bool tag decides whether the distance dependence is read from a file.

Hamster.get_onsite — Method

onsite=true

The onsite tag switches on the use of an extra parameter set for onsite interactions. If false, onsite matrix elements are computed evaluating the distance-dependence function at r=0.

Hamster.get_orbitals — Method

orbitals=[]

The orbitals tag specifies the set of orbitals used as the basis for the ion of the given type. If left empty, no orbitals are assigned and the corresponding atomic species is excluded from the electronic structure.

Hamster.get_rllm_file — Method

rllm_file=rllm.dat

The rllm_file::String tag sets the name of the file where the distance dependence is stored.

Hamster.get_sepNN — Method

sepNN=false

The sepNN::Bool tag switches on the use of an extra parameter set for nearest-neighbor interactions (compared to further away interactions). This can improve accuracy, however, may at the same time affect transferability negatively.

Hamster.get_tmethod — Method

tmethod=rotation

The tmethod::String tag selects the method used to compute the Slater-Koster transformation matrix. In practice, this choice should not affect the results, although rotation is slightly more efficient.

Possible options:

rotation: constructs the new reference system from the rotation matrix (default).

gramschmidt: constructs the new reference system using a Gram-Schmidt procedure.

Hamster.SparsityMode — Type

Abstract type SparsityMode serves as a base for indicating the sparsity of matrices. These types enable dispatching based on the sparsity mode.

Subtypes

Sparse <: SparsityMode: Represents matrices with sparse storage.

Dense <: SparsityMode: Represents matrices with dense storage.

Hamster.get_sp_diag — Method

sp_diag=false

The sp_diag::Bool tag switches between dense and sparse methods for matrix diagonalization (only affects Hᵏ and diagonalization). This can not be combined with optimization.

Hamster.get_sp_mode — Method

sp_mode=false

The sp_mode::Bool tag switches between dense and sparse matrix methods. This will only affect the computation of Hᴿ (not Hᵏ or diagonalization) and gradient computations when doing optimization.

Hamster.get_sp_tol — Method

sp_tol=1e-10

The sp_tol::Float tag sets a tolerance for values to be considered zero.

Hamster.get_tb_model — Method

tb_model=true

The tb_model tag switches on the use of a TB model in the effective Hamiltonian model.