General Options

The general options are placed inside an Options block:

begin Options
    ...
end

Hamster.get_nhamster — Method

nhamster=1

The nhamster tag sets the number of Hamster processes to be spawned for parallel tasks.

Hamster.get_nthreads_bands — Method

nthreads_bands=JULIANUMTHREADS

The nthreads_bands::Int tag sets the number of tasks to work on energy bands simultaneously.

Hamster.get_nthreads_blas — Method

nthreads_blas=1

The nthreads_blas::Int tag sets the number of threads used by the BLAS library.

Hamster.get_nthreads_kpoints — Method

nthreads_kpoints=JULIANUMTHREADS

The nthreads_kpoints::Int tag sets the number of tasks to work on kpoints simultaneously.

Hamster.set_seed! — Method

seed=none

The seed tag can be used to set a custom seed for RNG.

Hamster.get_orbitals — Method

orbitals=[]

The orbitals tag specifies the set of orbitals used as the basis for the ion of the given type. If left empty, no orbitals are assigned and the corresponding atomic species is excluded from the electronic structure.

Hamster.get_rllm_file — Method

rllm_file=rllm.dat

The rllm_file::String tag sets the name of the file where the distance dependence is stored.

Hamster.get_tmethod — Method

tmethod=rotation

The tmethod::String tag selects the method used to compute the Slater-Koster transformation matrix. In practice, this choice should not affect the results, although rotation is slightly more efficient.

Possible options:

rotation: constructs the new reference system from the rotation matrix (default).

gramschmidt: constructs the new reference system using a Gram-Schmidt procedure.

Hamster.SparsityMode — Type

Abstract type SparsityMode serves as a base for indicating the sparsity of matrices. These types enable dispatching based on the sparsity mode.

Subtypes

Sparse <: SparsityMode: Represents matrices with sparse storage.

Dense <: SparsityMode: Represents matrices with dense storage.

Hamster.get_sp_diag — Method

sp_diag=false

The sp_diag::Bool tag switches between dense and sparse methods for matrix diagonalization (only affects Hᵏ and diagonalization). This can not be combined with optimization.

Hamster.get_sp_mode — Method

sp_mode=false

The sp_mode::Bool tag switches between dense and sparse matrix methods. This will only affect the computation of Hᴿ (not Hᵏ or diagonalization) and gradient computations when doing optimization.